[allocator] register oom observers on every device
This change is to match the behavior of _record_memory_history which was
recently changed to enable history recording on all devices rather than
the current one. It prevents confusing situations where the observer
was registered before the device was set for the training run.

It also ensures the Python bindings initialize cuda so that the allocators
have been created before registering the callbacks.

ghstack-source-id: 8a72b8cd29cb075dcf31a66ba5c32e381c2ddfd7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107399